RBAC for SRS API

+++ api/src/main/java/org/ihtsdo/buildcloud/controller/BuildController.java
@@ -1,20 +1,15 @@
 package org.ihtsdo.buildcloud.controller;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
+import com.mangofactory.swagger.annotations.ApiIgnore;
+import com.wordnik.swagger.annotations.Api;
+import com.wordnik.swagger.annotations.ApiOperation;
 import org.ihtsdo.buildcloud.controller.helper.HypermediaGenerator;
 import org.ihtsdo.buildcloud.entity.Build;
 import org.ihtsdo.buildcloud.entity.BuildConfiguration;
 import org.ihtsdo.buildcloud.entity.QATestConfig;
-import org.ihtsdo.buildcloud.entity.User;
+import org.ihtsdo.buildcloud.security.IsAuthenticatedAsAdmin;
+import org.ihtsdo.buildcloud.security.IsAuthenticatedAsAdminOrReleaseManager;
+import org.ihtsdo.buildcloud.security.IsAuthenticatedAsAdminOrReleaseManagerOrUser;
 import org.ihtsdo.buildcloud.service.BuildService;
 import org.ihtsdo.buildcloud.service.PublishService;
 import org.ihtsdo.buildcloud.service.helper.ProcessingStatus;
@@ -28,16 +23,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StreamUtils;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
-import com.mangofactory.swagger.annotations.ApiIgnore;
-import com.wordnik.swagger.annotations.Api;
-import com.wordnik.swagger.annotations.ApiOperation;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequestMapping("/centers/{releaseCenterKey}/products/{productKey}/builds")
@@ -58,19 +55,21 @@ public class BuildController {
 	private static final String[] BUILD_LINKS = {"configuration","qaTestConfig", "inputfiles","inputGatherReport", "inputPrepareReport", "outputfiles", "buildReport", "logs", "buildLogs", "preConditionCheckReports", "postConditionCheckReports", "classificationResultsOutputFiles"};
 
 	@RequestMapping( method = RequestMethod.POST )
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ApiOperation( value = "Create a build",
 		notes = "Create a build for given product key and release center key and returns build id" )
 	@ResponseBody
 	public ResponseEntity<Map<String, Object>> createBuild(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
 			final HttpServletRequest request) throws BusinessServiceException {
 		final String currentUser = SecurityUtil.getUsername();
-		final Build build = buildService.createBuildFromProduct(releaseCenterKey, productKey, null, currentUser != null ? currentUser : User.ANONYMOUS_USER, null, null, null, null, null);
+		final Build build = buildService.createBuildFromProduct(releaseCenterKey, productKey, null, currentUser, null, null, null, null, null);
 
 		final boolean currentResource = false;
 		return new ResponseEntity<>(hypermediaGenerator.getEntityHypermedia(build, currentResource, request, BUILD_LINKS), HttpStatus.CREATED);
 	}
 
 	@RequestMapping( value = "/{buildId}", method = RequestMethod.DELETE )
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ApiOperation( value = "Delete a build",
 			notes = "" +
 					"Delete a build for given product key and release center key and build id" )
@@ -81,12 +80,17 @@ public class BuildController {
 
 		ifBuildIsNullThrow(productKey, buildId, build);
 
+		if (!CollectionUtils.isEmpty(build.getTags()) && build.getTags().contains(Build.Tag.PUBLISHED)) {
+			throw new BusinessServiceException("You can not delete the PUBLISHED build");
+		}
+
 		buildService.delete(releaseCenterKey, productKey, buildId);
 
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
 	@RequestMapping( method = RequestMethod.GET )
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation( value = "Returns a list all builds for a logged in user",
 		notes = "Returns a list all builds visible to the currently logged in user, "
 			+ "so this could potentially span across Release Centres" )
@@ -100,6 +104,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}", method = RequestMethod.GET )
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Get a build id",
 	notes = "Returns a single build object for given key" )
@@ -114,6 +119,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/configuration", produces = "application/json", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves configuration details",
 		notes = "Retrieves configuration details for given product key, release center key, and build id" )
@@ -129,6 +135,7 @@ public class BuildController {
 	
 	
 	@RequestMapping(value = "/{buildId}/qaTestConfig", produces = "application/json", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves QA test configuration details",
 		notes = "Retrieves configuration details for given product key, release center key, and build id" )
@@ -144,6 +151,7 @@ public class BuildController {
 	
 	
 	@RequestMapping(value = "/{buildId}/buildReport", produces = "application/json", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves build report details",
 		notes = "Retrieves buildReport details for given product key, release center key, and build id" )
@@ -160,6 +168,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/inputfiles", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves list of input file names",
 		notes = "Retrieves list of input file names for given release center, product key and build id" )
@@ -171,6 +180,7 @@ public class BuildController {
 	}
 	
 	@RequestMapping(value = "/{buildId}/inputPrepareReport", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves the report for preparing input source files.",
 		notes = "Product key and build id are required. And the report might not exist if no preparation is required." )
@@ -187,6 +197,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/inputGatherReport", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves the report for gathering input source files.",
 			notes = "Product key and build id are required. And the report might not exist if no preparation is required." )
@@ -204,6 +215,7 @@ public class BuildController {
 
 
 	@RequestMapping(value = "/{buildId}/inputfiles/{inputFileName:.*}", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation( value = "Download a specific file",
 		notes = "Download a specific file content for given release center, product key, build id and given input file name combination" )
 	public void getPackageInputFile(@PathVariable final String releaseCenterKey, @PathVariable final String productKey, @PathVariable final String buildId,
@@ -216,6 +228,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/outputfiles", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves a list of file names from output directory",
 		notes = "Retrieves a list of file names from output directory for given release center, "
@@ -228,6 +241,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/outputfiles/{outputFileName:.*}", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation( value = "Download a specific file from output directory",
 		notes = "Download a specific file from output directory for given release center, "
 			+ "product key, build id and file name combination" )
@@ -241,6 +255,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/trigger", method = RequestMethod.POST)
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ResponseBody
 	@ApiIgnore
 	public Map<String, Object> triggerProduct(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
@@ -252,6 +267,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/visibility", method = RequestMethod.POST)
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ResponseBody
 	@ApiOperation( value = "Update visibility for build", notes = "Update an existing build with the visibility flag")
 	public ResponseEntity updateVisibility(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
@@ -261,6 +277,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/tags", method = RequestMethod.POST)
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ResponseBody
 	@ApiOperation( value = "Update tags for build", notes = "Update an existing build with the tags")
 	public ResponseEntity updateTags(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
@@ -271,6 +288,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/publish", method = RequestMethod.POST)
+	@IsAuthenticatedAsAdmin
 	@ResponseBody
 	@ApiOperation( value = "Publish a release for given build id",
 	notes = "Publish release for given build id to make it available in repository for wider usages" )
@@ -283,6 +301,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/publish/status", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdmin
 	@ResponseBody
 	@ApiOperation( value = "Get publishing release status",
 			notes = "Get publishing release status for given build id")
@@ -299,6 +318,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/logs" , method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves a list of build log file names",
 		notes = "Retrieves a list of build log file names for given release center, product key, and build id" )
@@ -309,6 +329,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/logs/{logFileName:.*}", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation( value = "Download a specific build log file",
 		notes = "Download a specific log file for given release center, "
 		+ "product key, build id and file name combination" )
@@ -325,6 +346,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/logs/{logFileName:.*}", method = RequestMethod.HEAD)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation(value = "Download a specific build log file", notes = "Download a specific log file for given release center, "
 			+ "product key, build id and file name combination")
 	public void getBuildLogHead(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
@@ -338,6 +360,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/preConditionCheckReports", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves Pre-Condition Check Report",
 			notes = "Retrieves configuration details for given product key, release center key, and build id" )
@@ -354,6 +377,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/postConditionCheckReports", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves Post-Condition Check Report",
 			notes = "Retrieves configuration details for given product key, release center key, and build id" )
@@ -370,6 +394,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/classificationResultsOutputFiles", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ResponseBody
 	@ApiOperation( value = "Retrieves classification results for output files",
 			notes = "Retrieves configuration details for given product key, release center key, and build id" )
@@ -380,6 +405,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/classificationResultsOutputFiles/{inputFileName:.*}", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation( value = "Download a specific file",
 			notes = "Download a specific file content for given release center, product key, build id and given file name combination" )
 	public void getClassificationResultsOutputFiles(@PathVariable final String releaseCenterKey, @PathVariable final String productKey, @PathVariable final String buildId,
@@ -408,6 +434,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/cancel", method = RequestMethod.POST)
+	@IsAuthenticatedAsAdminOrReleaseManager
 	@ApiOperation(value = "Cancel a running build job")
 	public void requestCancelBuild(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
 								   @PathVariable final String buildId, final HttpServletResponse response) throws ResourceNotFoundException, BadConfigurationException {
@@ -416,6 +443,7 @@ public class BuildController {
 	}
 
 	@RequestMapping(value = "/{buildId}/buildLogs", method = RequestMethod.GET)
+	@IsAuthenticatedAsAdminOrReleaseManagerOrUser
 	@ApiOperation(value = "Get the full logs of the build process")
 	public void getFullBuildLogs(@PathVariable final String releaseCenterKey, @PathVariable final String productKey,
 								   @PathVariable final String buildId, HttpServletResponse response) throws IOException {
